+++ lib/enums.dart
@@ -9,17 +9,6 @@ extension CameraToString on Camera {
   String get stringValue => _enumCaseToString(describeEnum(this));
 }
 
-enum FaceDetector {
-  auto,
-  classic,
-  mlKit,
-  blazeface,
-}
-
-extension FaceDetectorToString on FaceDetector {
-  String get stringValue => _enumCaseToString(describeEnum(this));
-}
-
 enum Orientation {
   portrait,
   landscape,
@@ -36,4 +25,4 @@ String _enumCaseToString(String text) {
   return text.replaceAllMapped(exp, (Match m) => ('_' + m.group(0)!)).toLowerCase();
 }
 
-enum Canceller { user, app }
+enum Canceler { user, app }